Pattern 컴포넌트의 로직을 PatternContainer 로 분리함
